Extract `dev` module
Adds a new `dev` module containing various integration test
functionality intended to be reusable across projects.
